test: extracted tests for delegates into separate file. Moved testing resources into subdirectories
