+++ compiler-plugin-test/src/test/resources/kotlin/delegates/IterationAlignDelegateNeighboringFor.kt
