+++ compiler-plugin-test/src/test/resources/kotlin/delegates/IterationAlignDelegateNeighboringForEach.kt
