+++ compiler-plugin-test/src/test/resources/kotlin/delegates/IterationDelegateAlignNeighboringFor.kt
