+++ compiler-plugin-test/src/test/resources/kotlin/delegates/IterationDelegateAlignNeighboringForEach.kt
