+++ compiler-plugin-test/src/test/resources/kotlin/delegates/IterationDelegateNeighboringFor.kt
