+++ compiler-plugin-test/src/test/resources/kotlin/delegates/IterationDelegateNeighboringForEach.kt
