+++ compiler-plugin-test/src/test/resources/kotlin/delegates/IterationDelegateWithNestedFunNeighboringFor.kt
