+++ compiler-plugin-test/src/test/resources/kotlin/delegates/IterationDelegateWithNestedFunNeighboringForEach.kt
