+++ compiler-plugin-test/src/test/resources/kotlin/delegates/IterationDelegatedNestedFunNeighboringFor.kt
