+++ compiler-plugin-test/src/test/resources/kotlin/delegates/IterationDelegatedNestedFunNeighboringForEach.kt
