+++ compiler-plugin-test/src/test/resources/kotlin/delegates/IterationRecursiveDelegateNeighboringFor.kt
