+++ compiler-plugin-test/src/test/resources/kotlin/delegates/IterationRecursiveDelegateNeighboringForEach.kt
