+++ compiler-plugin-test/src/test/resources/kotlin/simpleIterations/IterationAlignNeighboringFor.kt
