+++ compiler-plugin-test/src/test/resources/kotlin/simpleIterations/IterationAlignNeighboringForEach.kt
