+++ compiler-plugin-test/src/test/resources/kotlin/simpleIterations/IterationExtAlignNeighboringFor.kt
