+++ compiler-plugin-test/src/test/resources/kotlin/simpleIterations/IterationExtAlignNeighboringForEach.kt
