+++ compiler-plugin-test/src/test/resources/kotlin/simpleIterations/IterationNeighboringFor.kt
