+++ compiler-plugin-test/src/test/resources/kotlin/simpleIterations/IterationNeighboringForEach.kt
