+++ compiler-plugin-test/src/test/resources/kotlin/simpleIterations/IterationWithNestedFunNeighboringFor.kt
