+++ compiler-plugin-test/src/test/resources/kotlin/simpleIterations/IterationWithNestedFunNeighboringForEach.kt
