+++ compiler-plugin-test/src/test/resources/kotlin/simpleIterations/OutsideAggregateNeighboringFor.kt
