+++ compiler-plugin-test/src/test/resources/kotlin/simpleIterations/OutsideAggregateNeighboringForEach.kt
